+++ lib/uploadcare/entity/uploader.rb
@@ -33,7 +33,13 @@ module Uploadcare
       # upload single file
       def self.upload_file(file, options = {})
         response = UploaderClient.new.upload_many([file], options)
-        Uploadcare::Entity::File.info(response.success.to_a.flatten[-1])
+        uuid = response.success.values.first
+        if Uploadcare.config.secret_key.nil?
+          Uploadcare::Entity::File.new(file_info(uuid).success)
+        else
+          # we can get more info about the uploaded file
+          Uploadcare::Entity::File.info(uuid)
+        end
       end
 
       # upload multiple files
@@ -61,6 +67,12 @@ module Uploadcare
       # @param url [String]
       def self.get_upload_from_url_status(token)
         UploaderClient.new.get_upload_from_url_status(token)
+      end
+
+      # Get information about an uploaded file (without the secret key)
+      # @param uuid [String]
+      def self.file_info(uuid)
+        UploaderClient.new.file_info(uuid)
       end
 
       class << self